# Reduce spread operators to improve runtime performance