--- a/library/CHANGELOG.md
+++ b/library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `getPipeInfo` util and refactor code for performance reasons (pull request #46)
+
 ## v0.12.0 (August 11, 2023)
 
 - Change input type of `mimeType` validation to `Blob`