--- a/library/src/schemas/bigint/bigint.ts
+++ b/library/src/schemas/bigint/bigint.ts
@@ -1,6 +1,10 @@
 import { ValiError } from '../../error/index.ts';
 import type { BaseSchema, Pipe } from '../../types.ts';
-import { executePipe, getErrorAndPipe } from '../../utils/index.ts';
+import {
+  executePipe,
+  getErrorAndPipe,
+  getPipeInfo,
+} from '../../utils/index.ts';
 
 /**
  * Bigint schema type.
@@ -71,7 +75,7 @@ export function bigint(
       }
 
       // Execute pipe and return output
-      return executePipe(input, pipe, { ...info, reason: 'bigint' });
+      return executePipe(input, pipe, getPipeInfo(info, 'bigint'));
     },
   };
 }