--- a/library/src/schemas/blob/blobAsync.ts
+++ b/library/src/schemas/blob/blobAsync.ts
@@ -1,6 +1,10 @@
 import { ValiError } from '../../error/index.ts';
 import type { BaseSchemaAsync, PipeAsync } from '../../types.ts';
-import { executePipeAsync, getErrorAndPipe } from '../../utils/index.ts';
+import {
+  executePipeAsync,
+  getErrorAndPipe,
+  getPipeInfo,
+} from '../../utils/index.ts';
 
 /**
  * Blob schema async type.
@@ -74,7 +78,7 @@ export function blobAsync(
       }
 
       // Execute pipe and return output
-      return executePipeAsync(input, pipe, { ...info, reason: 'blob' });
+      return executePipeAsync(input, pipe, getPipeInfo(info, 'blob'));
     },
   };
 }