--- a/library/src/utils/getPipeInfo/index.ts
+++ b/library/src/utils/getPipeInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './getPipeInfo.ts';