I found that removing the object spread operator greatly improve the runtime performance of `parse`. Without this PR (`main`: 506149a2de9755f91a033f340a42917b8457c82e), I confirmed that in many benchmarks, Valibot is much slower than Zod but after this PR, Valibot is mostly faster than Zod.

I created this PR so that people can run benchmarks and we can discuss the actual implementation details to move forward.

## Benchmarks
I used [this benchmark](https://github.com/fabian-hiller/valibot/issues/19#issue-1824216605) to confirm this PR makes a big impact at least for my environment (MacBook Air M2). I would appreciate if more people can confirm this. Note that I removed yup and Joi since they seemed to affect the benchmark results. We should make sure the benchmark is actually correct but I think **it's safe to say that Valibot can be actually faster than Zod with small tweaks like this PR**

- Before this PR (`main`: 506149a2de9755f91a033f340a42917b8457c82e)
```
: zod validate x 2,258,349 ops/sec ±1.14% (98 runs sampled)
: valibot validate x 713,707 ops/sec ±0.48% (93 runs sampled)
Fastest is zod validate
```

- After this PR (97ebba69c1cda0edabe9a4086054da3d1939edbf)
```
: zod validate x 2,289,850 ops/sec ±0.72% (101 runs sampled)
: valibot validate x 2,762,403 ops/sec ±0.07% (98 runs sampled)
Fastest is valibot validate
```

Note that I also tested this in https://github.com/moltar/typescript-runtime-type-benchmarks/pull/1112

## Backstory 
Recently I got amazed how Valibot is designed ❤️  ([Twitter thread in Japanese](https://twitter.com/naruaway/status/1687000165406904320)).

The implementation is really clean, modular, and simple. Since I am a huge fun of Zod and its bundle size is big, I was trying to invent special compiler for Zod to reduce the bundle size and it turned out that it was futile since Zod's implementation is hard to untangle and simplify.
So I just decided to use Valibot for my personal projects and I also want to support wider adoption for more serious production usage as well.

However, if we think of Valibot as "Zod replacement", runtime performance of `parse` is definitely a concern. For example, if we implement some validation logic in server side, we might suffer from DoS attack more often than when using Zod if the validation performance of Valibot is slower. Actually I expected Valibot is faster since the implementation is simpler than Zod. But it turned out that it's much slower than Zod in several benchmarks ([one example](https://github.com/fabian-hiller/valibot/issues/19#issue-1824216605)). I also tested several patterns locally and Zod was always much faster. After I played with Valibot codebase, I just found that we use the object spread operator in the hotpath of `parse` and removing that greatly improves the performance in my environment.


## Misc
I also tested some other things like caching object keys in `object.ts` and it looks like it's also improving the performance in some benchmarks. But I want to make sure we have good suite of benchmarks before discussing further on this. Reducing the object spread operator seems to have the biggest impact.