@fabian-hiller 2023-08-03
No, I think that Valibot already has a good foundation. I don't expect any big changes at the moment. If I make better progress with my bachelor thesis than expected, I might merge this PR much earlier.