@naruaway 2023-08-04
@fabian-hiller 
> No, I think that Valibot already has a good foundation.
Good to hear because my impression was the same! Clean and simple, small codebase ❤️ 

Let me actually try to build extensive benchmark suite so that we can make sure Valibot is always fast in any environment. I'll create another issue to track benchmarking in general later