@naruaway 2023-08-07
Now I created tracking issue for performance improvement in general https://github.com/fabian-hiller/valibot/issues/73 so we can use this PR only for the specific change