@fabian-hiller 2023-08-12
I have now done this investigation myself. You are right. The impact on performance is extreme. In my investigation, I noticed that the following makes no difference to performance. Therefore, I prefer the variant that requires less code.

```ts
// With ternary operator
export const mergeInfoWithReason = (info: ParseInfo | undefined, reason: IssueReason) => {
  return info ? {
    origin: info.origin,
    path: info.path,
    abortEarly: info.abortEarly,
    abortPipeEarly: info.abortPipeEarly,
    reason,
  } : { reason };
}

// Without ternary operator
export const mergeInfoWithReason = (info: ParseInfo | undefined, reason: IssueReason) => {
  return {
    origin: info?.origin,
    path: info?.path,
    abortEarly: info?.abortEarly,
    abortPipeEarly: info?.abortPipeEarly,
    reason,
  };
}
```

I appreciate your research and would leave it up to you to implement this change for the entire library. Otherwise, I can implement it and mention you in the release notes. Before we go into the implementation, I would clarify a few details about the final code with you.