@naruaway 2023-08-12
@fabian-hiller thanks for following up on this!
Now I am finishing up an improving of https://github.com/naruaway/valibot-benchmarks to be able to leave an evidence in reproducible way since having reliable benchmarks is critical for any performance improvements. I will restart working on this tomorrow evening  👍 

Yeah I would aim for minimal code to get the same level of perfs across major runtimes (combinations of (macOS (ARM), Linux (x86)) x (Bun(JavaScriptCore) and Node.js/Deno (V8)))