@fabian-hiller 2023-08-12
> I appreciate your research and would leave it up to you to implement this change for the entire library. Otherwise, I can implement it and mention you in the release notes. Before we go into the implementation, I would clarify a few details about the final code with you.

Do you want to make the changes, or should I do it? If so, in what time frame can you do it?

I'm reviewing the performance today and tomorrow as part of my thesis. I want to implement any improvements over the weekend to be able to focus on other areas of the paper next week.