@naruaway 2023-08-12
@fabian-hiller Ok, I was planning to finish up benchmark suite thing first but I anyway we are on the same page that this is significant perf booster, let me do the change now (I have time now before sleeping) 👍 I might follow up precise perf results / small tweaks later then