@fabian-hiller 2023-08-12
Great! Can you rename `mergeInfoWithReason` to `getPipeInfo` and use `function` instead of an arrow function?