@naruaway 2023-08-12
@fabian-hiller I just realized that we might change `executePipe` itself to accept `ParseInfo` and `IssueReason` separately but just to get quick perf boost without further churn, let me do the simplest change (`getPipeInfo` function everywhere) first and potentially file another PR