@fabian-hiller 2023-08-12
You're right. The logic can also be in the pipe function.