@naruaway 2023-08-12
@fabian-hiller, anyway I pushed `getPipeInfo` change and I believe I did not miss any other places:
```
# On main branch
$ rg -l '...info, reason'|wc -l
23

# On this PR
$ rg -l '...info, reason'|wc -l
0
```
Could you review and merge if it looks good? (of course feel free to tweak or point out anything 👍 )

Note that this is not enough to "beat" Zod in benchmarks for valid data most of the time. We also need to eliminate the other spread operator like [this](https://github.com/fabian-hiller/valibot/blob/ce49b0b4809a62035767c4b7697fabddd871392c/library/src/schemas/object/object.ts#L114-L117). And this cannot be put in executePipe. So let me create similar util for now. I'll add the change in this PR if you do not merge this PR yet. If you merge this PR, I'll file another PR