@fabian-hiller 2023-08-12
I'll check it in 30 minutes and give you feedback. Feel free to create further PRs for the other spread operators. I'll continue later with my test as well.