@naruaway 2023-08-12
Ah I missed  multi line patterns:
```
...info,
reason: 'tuple',
```
Thanks for the follow up and refactor 👍